cleaning up old utils class
